+++ Tests/Assets/TestModules/_Plugz.psm1
@@ -1,0 +1,16 @@
+function _Export-PlugzProfile{
+    _Get-PlugzConfig
+}
+function _Get-PlugzConfig{
+    _Import-Configuration
+}
+function _Import-Plugz{
+    _Get-PlugzConfig
+    _Test-CalledFromProfile
+}
+function _Save-PlugzConfig{
+    _Export-Configuration
+}
+function _Test-CalledFromProfile{
+}
+