+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# FindFunctionCalls
+# WhoCalled
+
+## 2.0.0
+
+- Rename module `FindFunctionCalls` => `WhoCalled`
+- Rename `Find-FunctionCall` => `Find-Call`
+- Add `Find-Caller`
 
 ## 1.4.5